refactor levels, instructions into diff pages
