+++ Webapp/client/src/App.js
@@ -1,9 +1,11 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { Container } from "react-bootstrap";
 import NavBar from "./NavBar";
-import Leaderboard from "./Leaderboard";
-import History from "./History";
-import Home from "./Home";
+import Leaderboard from "./LeaderboardComponents/Leaderboard";
+import History from "./HistoryComponents/History";
+import Home from "./HomeComponents/Home";
+import Levels from "./LevelsComponents/Levels";
+import Instructions from "./InstructionsComponents/Instructions";
 import "./App.css";
 
 function App() {
@@ -15,6 +17,8 @@ function App() {
           <Routes>
             <Route path="/history/:username" element={<History />} />
             <Route path="/leaderboard" element={<Leaderboard />} />
+            <Route path="/levels" element={<Levels />} />
+            <Route path="/instructions" element={<Instructions />} />
             <Route path="/" element={<Home />} />
           </Routes>
         </BrowserRouter>
