+++ Webapp/client/src/InstructionsComponents/Instructions.css
@@ -1,4 +1,4 @@
-.home .react-bootstrap-table table {
+.instructions .react-bootstrap-table table {
   table-layout: auto !important;
   max-width: 600px;
 }
